pyproject.toml
[project]
name = "lyrics-llm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit==1.43.2",
    "llama-index==0.12.25",
    "llama-index-llms-ollama==0.5.3",
    "llama-index-embeddings-ollama==0.6.0",
    "nltk==3.9.1",
    "PyGithub==2.6.1",
    "python-dotenv==1.0.1",
]
